Create a Private repository to hold custom modules
- Create a /Private directory
- Add a ls module and migrate ls response code to function
    - ls response code now calls function
- Add local ls module as a dependency
