+++ package.json
@@ -26,7 +26,6 @@
   "dependencies": {
     "eslint-config-google": "^0.13.0",
     "express": "^4.17.1",
-    "file-system": "^2.2.2",
-    "path": "^0.12.7"
+    "ls": "file:Private/ls"
   }
 }
